+++ end_to_end_tests/golden-record-custom/my_test_api_client/api/tests/no_response_tests_no_response_get.py
@@ -1,0 +1,25 @@
+import httpx
+
+Client = httpx.Client
+
+
+def _build_response(*, response: httpx.Response) -> httpx.Response[None]:
+    return httpx.Response(
+        status_code=response.status_code,
+        content=response.content,
+        headers=response.headers,
+        parsed=None,
+    )
+
+
+def httpx_request(
+    *,
+    client: Client,
+) -> httpx.Response[None]:
+
+    response = client.request(
+        "get",
+        "/tests/no_response",
+    )
+
+    return _build_response(response=response)