+++ openapi_python_client/cli.py
@@ -96,10 +96,20 @@ def handle_errors(errors: Sequence[GeneratorError]) -> None:
         raise typer.Exit(code=1)
 
 
+custom_template_path_options = {
+    "help": "A path to a directory containing custom template(s)",
+    "file_okay": False,
+    "dir_okay": True,
+    "readable": True,
+    "resolve_path": True,
+}
+
+
 @app.command()
 def generate(
     url: Optional[str] = typer.Option(None, help="A URL to read the JSON from"),
     path: Optional[pathlib.Path] = typer.Option(None, help="A path to the JSON file"),
+    custom_template_path: Optional[pathlib.Path] = typer.Option(None, **custom_template_path_options),  # type: ignore
 ) -> None:
     """ Generate a new OpenAPI Client library """
     from . import create_new_client
@@ -110,7 +120,7 @@ def generate(
     if url and path:
         typer.secho("Provide either --url or --path, not both", fg=typer.colors.RED)
         raise typer.Exit(code=1)
-    errors = create_new_client(url=url, path=path)
+    errors = create_new_client(url=url, path=path, custom_template_path=custom_template_path)
     handle_errors(errors)
 
 
@@ -118,6 +128,7 @@ def generate(
 def update(
     url: Optional[str] = typer.Option(None, help="A URL to read the JSON from"),
     path: Optional[pathlib.Path] = typer.Option(None, help="A path to the JSON file"),
+    custom_template_path: Optional[pathlib.Path] = typer.Option(None, **custom_template_path_options),  # type: ignore
 ) -> None:
     """ Update an existing OpenAPI Client library """
     from . import update_existing_client
@@ -129,5 +140,5 @@ def update(
         typer.secho("Provide either --url or --path, not both", fg=typer.colors.RED)
         raise typer.Exit(code=1)
 
-    errors = update_existing_client(url=url, path=path)
+    errors = update_existing_client(url=url, path=path, custom_template_path=custom_template_path)
     handle_errors(errors)